+++ Challenge4/Challenge4/OutputWriter.cs
@@ -5,13 +5,13 @@ namespace Challenge4
 {
     public class OutputWriter : IOutputWriter
     {
-        public void WriteOutput(List<(int, int)> candiesAverages, string outputPath)
+        public void WriteOutput(List<(long, long)> candiesAverages, string outputPath)
         {
             var lines = BuildOutputLines(candiesAverages);
             File.WriteAllLines(outputPath, lines);
         }
 
-        private IEnumerable<string> BuildOutputLines(List<(int Candies, int People)> candiesAverages)
+        private IEnumerable<string> BuildOutputLines(List<(long Candies, long People)> candiesAverages)
         {
             for (var i = 0; i < candiesAverages.Count; i++)
             {
